Monte Carlo Simulation for Tic Tac Toe Bot
Code for the monte carlo simulation of a tic tac toe bot that optimizes itself to pick the next best move in a game of tic tac toe. The instructors in the course created the tic tac toe class and GUI in separate scripts and in this project both the tic tac toe class and GUI are imported to allow students to test their code for the monte carlo simulation. Students do not have access to the code that is being imported for this project for the GUI or the tic tac toe class.